split out verifier, fix DerivedKeyDelegatedSignatureService and test
