+++ xrpl4j-crypto-parent/xrpl4j-crypto-core/src/test/java/org/xrpl/xrpl4j/crypto/core/signing/AbstractDelegatedSignatureServiceTest.java
@@ -2,36 +2,22 @@ package org.xrpl.xrpl4j.crypto.core.signing;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.verifyNoMoreInteractions;
 import static org.mockito.Mockito.when;
+import static org.mockito.MockitoAnnotations.openMocks;
 
-import org.assertj.core.util.Sets;
-import org.junit.jupiter.api.Assertions;
+import com.google.common.collect.Sets;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
-import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
-import org.xrpl.xrpl4j.codec.addresses.UnsignedByteArray;
-import org.xrpl.xrpl4j.codec.addresses.VersionType;
-import org.xrpl.xrpl4j.crypto.core.AddressUtils;
 import org.xrpl.xrpl4j.crypto.core.KeyMetadata;
-import org.xrpl.xrpl4j.crypto.core.KeyStoreType;
-import org.xrpl.xrpl4j.crypto.core.keys.Ed25519KeyPairService;
-import org.xrpl.xrpl4j.crypto.core.keys.KeyPair;
-import org.xrpl.xrpl4j.crypto.core.keys.Passphrase;
 import org.xrpl.xrpl4j.crypto.core.keys.PublicKey;
-import org.xrpl.xrpl4j.crypto.core.keys.Secp256k1KeyPairService;
-import org.xrpl.xrpl4j.crypto.core.keys.Seed;
 import org.xrpl.xrpl4j.model.client.channels.UnsignedClaim;
-import org.xrpl.xrpl4j.model.transactions.Address;
 import org.xrpl.xrpl4j.model.transactions.Transaction;
 
-import java.util.Set;
-import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.HashSet;
 
 /**
  * Unit tests for {@link AbstractDelegatedSignatureService}.
@@ -39,389 +25,93 @@ import java.util.concurrent.atomic.AtomicBoolean;
 class AbstractDelegatedSignatureServiceTest {
 
   @Mock
-  private SignatureUtils signatureUtilsMock;
+  DelegatedTransactionSigner delegatedTransactionSignerMock;
+
   @Mock
-  private Signature ed25519SignatureMock;
+  DelegatedTransactionVerifier delegatedTransactionVerifierMock;
+
   @Mock
-  private Signature secp256k1SignatureMock;
+  KeyMetadata keyMetadataMock;
+
   @Mock
-  private KeyMetadata keyMetadataMock;
-  @Mock
-  private Transaction transactionMock;
-  @Mock
-  private AddressUtils addressServiceMock;
+  PublicKey publicKeyMock;
+
   @Mock
   private SignatureWithKeyMetadata signatureWithKeyMetaMock;
 
-  private KeyPair ed25519KeyPair;
-  private Address ed25519SignerAddress;
+  @Mock
+  private Signature signatureMock;
 
-  private KeyPair secp256k1KeyPair;
-  private Address secp256k1SignerAddress;
+  @Mock
+  private Transaction transactionMock;
 
-  private AtomicBoolean ed25519VerifyCalled;
-  private AtomicBoolean secp256k1VerifyCalled;
-  private VersionType keyType;
+  @Mock
+  private SingleSingedTransaction<Transaction> singleSignedTransactionMock;
 
-  private AbstractDelegatedSignatureService signatureService;
+  private AbstractDelegatedSignatureService delegatedSignatureService;
 
   @BeforeEach
   void setUp() {
-    MockitoAnnotations.openMocks(this);
-    ed25519VerifyCalled = new AtomicBoolean(false);
-    secp256k1VerifyCalled = new AtomicBoolean(false);
+    openMocks(this);
 
-    AddressUtils addressService = AddressUtils.getInstance();
-    Ed25519KeyPairService ed25519KeyPairService = Ed25519KeyPairService.getInstance();
-    this.ed25519KeyPair = ed25519KeyPairService.deriveKeyPair(
-      Seed.ed25519SeedFromPassphrase(Passphrase.of("hello"))
-    );
-    this.ed25519SignerAddress = addressService.deriveAddress(ed25519KeyPair.publicKey());
+    when(delegatedTransactionSignerMock.getPublicKey(keyMetadataMock)).thenReturn(publicKeyMock);
+    when(delegatedTransactionSignerMock.sign(keyMetadataMock, transactionMock)).thenReturn(singleSignedTransactionMock);
+    when(delegatedTransactionSignerMock.sign(eq(keyMetadataMock), any(UnsignedClaim.class))).thenReturn(signatureMock);
+    when(delegatedTransactionSignerMock.multiSign(keyMetadataMock, transactionMock))
+      .thenReturn(signatureWithKeyMetaMock);
 
-    Secp256k1KeyPairService secp256k1KeyPairService = Secp256k1KeyPairService.getInstance();
-    this.secp256k1KeyPair = secp256k1KeyPairService.deriveKeyPair(
-      Seed.secp256k1SeedFromPassphrase(Passphrase.of("hello"))
-    );
-    this.secp256k1SignerAddress = addressService.deriveAddress(secp256k1KeyPair.publicKey());
-
-    when(addressServiceMock.deriveAddress(any())).thenReturn(ed25519SignerAddress);
-
-    when(signatureUtilsMock.toSignableBytes(Mockito.<Transaction>any())).thenReturn(UnsignedByteArray.empty());
-    when(signatureUtilsMock.toMultiSignableBytes(any(), any())).thenReturn(UnsignedByteArray.empty());
-
-    when(signatureWithKeyMetaMock.signingKeyMetadata()).thenReturn(keyMetadataMock);
-    when(signatureWithKeyMetaMock.transactionSignature()).thenReturn(ed25519SignatureMock);
-
-    this.signatureService = new AbstractDelegatedSignatureService(
-      signatureUtilsMock,
-      addressServiceMock
-    ) {
-      @Override
-      public PublicKey getPublicKey(KeyMetadata keyMetadata) {
-        return keyType == VersionType.ED25519 ? ed25519KeyPair.publicKey() : secp256k1KeyPair.publicKey();
-      }
-
-      @Override
-      protected Signature edDsaSign(KeyMetadata privateKeyMetadata, UnsignedByteArray signableTransactionBytes) {
-        return ed25519SignatureMock;
-      }
-
-      @Override
-      protected Signature ecDsaSign(KeyMetadata privateKeyMetadata, UnsignedByteArray signableTransactionBytes) {
-        return secp256k1SignatureMock;
-      }
-
-      @Override
-      protected boolean edDsaVerify(
-        KeyMetadata keyMetadata, UnsignedByteArray signableTransactionBytes, Signature transactionSignature
-      ) {
-        ed25519VerifyCalled.set(true);
-        return ed25519VerifyCalled.get();
-      }
-
-      @Override
-      protected boolean ecDsaVerify(KeyMetadata keyMetadata, UnsignedByteArray signableTransactionBytes,
-        Signature transactionSignature) {
-        secp256k1VerifyCalled.set(true);
-        return secp256k1VerifyCalled.get();
-      }
-    };
-  }
-
-  ///////////////////
-  // Sign (Transaction)
-  ///////////////////
-
-  @Test
-  void signWithNullMetadata() {
-    Assertions.assertThrows(NullPointerException.class, () -> signatureService.sign(null, transactionMock));
+    this.delegatedSignatureService = new AbstractDelegatedSignatureService(
+      delegatedTransactionSignerMock,
+      delegatedTransactionVerifierMock
+    ) { };
   }
 
   @Test
-  void signWithNullTransaction() {
-    Assertions.assertThrows(NullPointerException.class,
-      () -> signatureService.sign(keyMetadataMock, (Transaction) null));
+  void getPublicKey() {
+    PublicKey actual = delegatedSignatureService.getPublicKey(keyMetadataMock);
+    assertThat(actual).isEqualTo(publicKeyMock);
   }
 
   @Test
-  void signEd25519() {
-    keyType = VersionType.ED25519;
-    signatureService.sign(keyMetadataMock, transactionMock);
-
-    verify(signatureUtilsMock).toSignableBytes(transactionMock);
-    verify(signatureUtilsMock).addSignatureToTransaction(transactionMock, ed25519SignatureMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
+  void signTransaction() {
+    SingleSingedTransaction<Transaction> actual = delegatedSignatureService.sign(keyMetadataMock, transactionMock);
+    assertThat(actual).isEqualTo(singleSignedTransactionMock);
+    verify(delegatedTransactionSignerMock).sign(keyMetadataMock, transactionMock);
   }
 
   @Test
-  void signSecp256k1() {
-    keyType = VersionType.SECP256K1;
-
-    signatureService.sign(keyMetadataMock, transactionMock);
-
-    verify(signatureUtilsMock).toSignableBytes(transactionMock);
-    verify(signatureUtilsMock).addSignatureToTransaction(transactionMock, secp256k1SignatureMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // Sign (UnsignedClaim)
-  ///////////////////
-
-  @Test
-  void signUnsignedClaimWithNullMetadata() {
-    UnsignedClaim unsignedClaimMock = mock(UnsignedClaim.class);
-    Assertions.assertThrows(NullPointerException.class, () -> signatureService.sign(null, unsignedClaimMock));
+  void signClaim() {
+    UnsignedClaim mockUnsignedClaim = mock(UnsignedClaim.class);
+    Signature actual = delegatedSignatureService.sign(keyMetadataMock, mockUnsignedClaim);
+    assertThat(actual).isEqualTo(signatureMock);
+    verify(delegatedTransactionSignerMock).sign(keyMetadataMock, mockUnsignedClaim);
   }
 
   @Test
-  void signUnsignedClaimWithNullTransaction() {
-    Assertions.assertThrows(NullPointerException.class,
-      () -> signatureService.sign(keyMetadataMock, (UnsignedClaim) null));
+  void multiSignTransaction() {
+    SignatureWithKeyMetadata actual = delegatedSignatureService.multiSign(keyMetadataMock, transactionMock);
+    assertThat(actual).isEqualTo(signatureWithKeyMetaMock);
+    verify(delegatedTransactionSignerMock).multiSign(keyMetadataMock, transactionMock);
   }
 
   @Test
-  void signUnsignedClaimEd25519() {
-    keyType = VersionType.ED25519;
-    UnsignedClaim unsignedClaimMock = mock(UnsignedClaim.class);
-    when(signatureUtilsMock.toSignableBytes(unsignedClaimMock)).thenReturn(UnsignedByteArray.empty());
-
-    Signature actual = signatureService.sign(keyMetadataMock, unsignedClaimMock);
-
-    verify(signatureUtilsMock).toSignableBytes(unsignedClaimMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-    assertThat(actual).isEqualTo(ed25519SignatureMock);
+  void verifySingleSigned() {
+    delegatedSignatureService.verifySingleSigned(signatureWithKeyMetaMock, transactionMock);
+    verify(delegatedTransactionVerifierMock).verifySingleSigned(signatureWithKeyMetaMock, transactionMock);
   }
 
   @Test
-  void signUnsignedClaimSecp256k1() {
-    keyType = VersionType.SECP256K1;
-    UnsignedClaim unsignedClaimMock = mock(UnsignedClaim.class);
-    when(signatureUtilsMock.toSignableBytes(unsignedClaimMock)).thenReturn(UnsignedByteArray.empty());
-
-    Signature actual = signatureService.sign(keyMetadataMock, unsignedClaimMock);
-
-    verify(signatureUtilsMock).toSignableBytes(unsignedClaimMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-    assertThat(actual).isEqualTo(secp256k1SignatureMock);
-  }
-
-  ///////////////////
-  // MultiSign
-  ///////////////////
-
-  @Test
-  void multiSignWithNullMetadata() {
-    Assertions.assertThrows(NullPointerException.class, () -> signatureService.multiSign(null, transactionMock));
+  void verifyMultiSignedWithoutMinSigners() {
+    HashSet<SignatureWithKeyMetadata> signatureSet = Sets.newHashSet(signatureWithKeyMetaMock);
+    delegatedSignatureService.verifyMultiSigned(signatureSet, transactionMock);
+    verify(delegatedTransactionVerifierMock).verifyMultiSigned(signatureSet, transactionMock);
   }
 
   @Test
-  void multiSignWithNullTransaction() {
-    Assertions.assertThrows(NullPointerException.class, () -> signatureService.multiSign(keyMetadataMock, null));
-  }
-
-  @Test
-  void multiSignEd25519() {
-    keyType = VersionType.ED25519;
-    when(addressServiceMock.deriveAddress(any())).thenReturn(ed25519SignerAddress);
-
-    SignatureWithKeyMetadata signatureWithKeyMetadata = signatureService.multiSign(
-      keyMetadataMock,
-      transactionMock
-    );
-    assertThat(signatureWithKeyMetadata.transactionSignature()).isEqualTo(ed25519SignatureMock);
-    assertThat(signatureWithKeyMetadata.signingKeyMetadata()).isEqualTo(keyMetadataMock);
-
-    verify(addressServiceMock).deriveAddress(any());
-    verifyNoMoreInteractions(addressServiceMock);
-    verify(signatureUtilsMock).toMultiSignableBytes(transactionMock, ed25519SignerAddress);
-    verify(signatureUtilsMock, times(0)).toSignableBytes(transactionMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  @Test
-  void multiSignSecp256k1() {
-    keyType = VersionType.SECP256K1;
-    when(addressServiceMock.deriveAddress(any())).thenReturn(secp256k1SignerAddress);
-
-    SignatureWithKeyMetadata signatureWithKeyMetadata = signatureService.multiSign(
-      keyMetadataMock,
-      transactionMock
-    );
-
-    assertThat(signatureWithKeyMetadata.transactionSignature()).isEqualTo(secp256k1SignatureMock);
-    assertThat(signatureWithKeyMetadata.signingKeyMetadata()).isEqualTo(keyMetadataMock);
-
-    verify(addressServiceMock).deriveAddress(any());
-    verifyNoMoreInteractions(addressServiceMock);
-    verify(signatureUtilsMock).toMultiSignableBytes(transactionMock, secp256k1SignerAddress);
-    verify(signatureUtilsMock, times(0)).toSignableBytes(transactionMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // verify (signatureWithKeyMetadata, transaction)
-  ///////////////////
-
-  @Test
-  void verifyWithNullMetadata() {
-    Assertions.assertThrows(NullPointerException.class,
-      () -> signatureService.verifySingleSigned(null, transactionMock));
-  }
-
-  @Test
-  void verifyWithNullTransaction() {
-    Assertions.assertThrows(NullPointerException.class,
-      () -> signatureService.verifySingleSigned(signatureWithKeyMetaMock, null));
-  }
-
-  @Test
-  void verifyEd25519() {
-    keyType = VersionType.ED25519;
-    boolean actual = signatureService.verifySingleSigned(signatureWithKeyMetaMock, transactionMock);
-
-    assertThat(actual).isTrue();
-    assertThat(ed25519VerifyCalled.get()).isTrue();
-    assertThat(secp256k1VerifyCalled.get()).isFalse();
-    verify(signatureUtilsMock).toSignableBytes(transactionMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  @Test
-  void verifySecp256k1() {
-    keyType = VersionType.SECP256K1;
-    boolean actual = signatureService.verifySingleSigned(signatureWithKeyMetaMock, transactionMock);
-
-    assertThat(actual).isTrue();
-    assertThat(secp256k1VerifyCalled.get()).isTrue();
-    assertThat(ed25519VerifyCalled.get()).isFalse();
-    verify(signatureUtilsMock).toSignableBytes(transactionMock);
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // verify (signatureWithKeyMetadata, transaction)
-  ///////////////////
-
-  @Test
-  void verifyMultiWithNullSet() {
-    Assertions.assertThrows(NullPointerException.class,
-      () -> signatureService.verifyMultiSigned(null, transactionMock));
-  }
-
-  @Test
-  void verifyMultiWithNullTransaction() {
-    Assertions.assertThrows(NullPointerException.class,
-      () -> signatureService.verifyMultiSigned(Sets.newHashSet(), null));
-  }
-
-  @Test
-  void verifyMultiWithEmptySet() {
-    Assertions.assertThrows(IllegalArgumentException.class,
-      () -> signatureService.verifyMultiSigned(Sets.newHashSet(), transactionMock));
-  }
-
-  @Test
-  void verifyMultiEd25519() {
-    keyType = VersionType.ED25519;
-    when(signatureWithKeyMetaMock.transactionSignature()).thenReturn(ed25519SignatureMock);
-
-    final Set<SignatureWithKeyMetadata> signatureWithKeyMetadataSet = Sets.newLinkedHashSet(signatureWithKeyMetaMock);
-    boolean actual = signatureService.verifyMultiSigned(signatureWithKeyMetadataSet, transactionMock, 1);
-
-    assertThat(actual).isTrue();
-    assertThat(ed25519VerifyCalled.get()).isTrue();
-    assertThat(secp256k1VerifyCalled.get()).isFalse();
-    verify(addressServiceMock).deriveAddress(ed25519KeyPair.publicKey());
-    verifyNoMoreInteractions(addressServiceMock);
-    verify(signatureUtilsMock).toMultiSignableBytes(any(), any());
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  @Test
-  void verifyMultiSecp256k1() {
-    keyType = VersionType.SECP256K1;
-    when(signatureWithKeyMetaMock.transactionSignature()).thenReturn(secp256k1SignatureMock);
-
-    final Set<SignatureWithKeyMetadata> signatureWithKeyMetadataSet = Sets.newLinkedHashSet(signatureWithKeyMetaMock);
-    boolean actual = signatureService.verifyMultiSigned(signatureWithKeyMetadataSet, transactionMock, 1);
-
-    assertThat(actual).isTrue();
-    assertThat(secp256k1VerifyCalled.get()).isTrue();
-    assertThat(ed25519VerifyCalled.get()).isFalse();
-    verify(addressServiceMock).deriveAddress(secp256k1KeyPair.publicKey());
-    verifyNoMoreInteractions(addressServiceMock);
-    verify(signatureUtilsMock).toMultiSignableBytes(any(), any());
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // EdDsaSign
-  ///////////////////
-
-  @Test
-  void edDsaSign() {
-    this.keyType = VersionType.ED25519;
-
-    Signature actual = signatureService.edDsaSign(keyMetadataMock,
-      UnsignedByteArray.empty());
-
-    assertThat(actual).isEqualTo(ed25519SignatureMock);
-    assertThat(ed25519VerifyCalled.get()).isFalse();
-    assertThat(secp256k1VerifyCalled.get()).isFalse();
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // EcDsaSign
-  ///////////////////
-
-  @Test
-  void ecDsaSign() {
-    this.keyType = VersionType.SECP256K1;
-
-    Signature actual = signatureService.ecDsaSign(keyMetadataMock, UnsignedByteArray.empty());
-
-    assertThat(actual).isEqualTo(secp256k1SignatureMock);
-
-    assertThat(secp256k1VerifyCalled.get()).isFalse();
-    assertThat(ed25519VerifyCalled.get()).isFalse();
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // edDsaVerify
-  ///////////////////
-
-  @Test
-  void edDsaVerify() {
-    this.keyType = VersionType.ED25519;
-
-    boolean actual = signatureService.edDsaVerify(
-      keyMetadataMock, UnsignedByteArray.empty(), ed25519SignatureMock
-    );
-
-    assertThat(actual).isTrue();
-    assertThat(ed25519VerifyCalled.get()).isTrue();
-    assertThat(secp256k1VerifyCalled.get()).isFalse();
-    verifyNoMoreInteractions(signatureUtilsMock);
-  }
-
-  ///////////////////
-  // ecDsaVerify
-  ///////////////////
-
-  @Test
-  void ecDsaVerify() {
-    this.keyType = VersionType.SECP256K1;
-
-    boolean actual = signatureService.ecDsaVerify(
-      keyMetadataMock, UnsignedByteArray.empty(), secp256k1SignatureMock
-    );
-
-    assertThat(actual).isTrue();
-    assertThat(secp256k1VerifyCalled.get()).isTrue();
-    assertThat(ed25519VerifyCalled.get()).isFalse();
-    verifyNoMoreInteractions(signatureUtilsMock);
+  void verifyMultiSignedWithMinSigners() {
+    HashSet<SignatureWithKeyMetadata> signatureSet = Sets.newHashSet(signatureWithKeyMetaMock);
+    int minSigners = 1;
+    delegatedSignatureService.verifyMultiSigned(signatureSet, transactionMock, minSigners);
+    verify(delegatedTransactionVerifierMock).verifyMultiSigned(signatureSet, transactionMock, minSigners);
   }
 }
